+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ## Unreleased
 
+- :bug: ![o] [#408](https://github.com/giraud/reasonml-idea-plugin/issues/408) let ... in not shown in let a = ... structure
 - :bug: ![o] [#406](https://github.com/giraud/reasonml-idea-plugin/issues/406) Faulty let open ... in parsing
 
 **NOTE**: minimal version supported is **2022.2**